fix: address issues outlined in #47
When Strict Mode is on, React will also run one extra setup+cleanup
cycle in development for every Effect. In this commit, we replace
the useState in favour of useRef to keep track of the necessary
state. We also introduce a clean-up function in the Effect to
mutate the Ref accordingly.